fix symredirect-host-build for macos

+++ symredirect-host-build.sh
@@ -3,17 +3,21 @@
 set -e
 set -x
 
-echo "*** Cloning ld64 repository ..."
-[ -d ld64 ] && rm -rf ld64
-git clone https://github.com/ProcursusTeam/ld64.git
+if [ $(uname -s) != "Darwin" ]; then
+    echo "*** Cloning ld64 repository ..."
+    [ -d ld64 ] && rm -rf ld64
+    git clone https://github.com/ProcursusTeam/ld64.git
 
-echo "*** Checking the non-darwin headers ..."
-if [ ! -d ./ld64/EXTERNAL_HEADERS/non-darwin ]; then
-    echo "The non-darwin headers are not present in the ld64 repository."
+    echo "*** Checking the non-darwin headers ..."
+    if [ ! -d ./ld64/EXTERNAL_HEADERS/non-darwin ]; then
+        echo "The non-darwin headers are not present in the ld64 repository."
+    fi
+
+    EXTRA_HEADER_FLAG="-I./ld64/EXTERNAL_HEADERS/non-darwin"
 fi
 
 echo "*** Building the symredirect-host executable ..."
-clang++ -v -std=c++11 -I./ld64/EXTERNAL_HEADERS/non-darwin -DVROOT_API_ALL -o symredirect-host symredirect.cpp
+clang++ -v -std=c++11 $EXTRA_HEADER_FLAG -o symredirect-host symredirect.cpp
 
 echo "*** Checking built symredirect-host executable ..."
